[Embedding] modify build rule for GPU EV
