+++ tensorflow/core/kernels/kv_variable_ops_gpu.h
@@ -17,7 +17,6 @@ limitations under the License.
 #define TENSORFLOW_CORE_KERNELS_KV_GPU_HASH_TABLE_H_
 
 #if GOOGLE_CUDA
-#if CUDA_ATOMIC
 
 #include "third_party/eigen3/unsupported/Eigen/CXX11/Tensor"
 #include "tensorflow/core/framework/op_kernel.h"
@@ -292,7 +291,6 @@ class EmbeddingVarGPU : public ResourceBase {
 
 }  // namespace tensorflow
 
-#endif  // CUDA_ATOMIC
 #endif  // GOOGLE_CUDA
 
 #endif  // TENSORFLOW_CORE_KERNELS_KV_GPU_HASH_TABLE_H_
