+++ tensorflow/core/kernels/training_ali_ops.cc
@@ -35,6 +35,10 @@ limitations under the License.
 #include "tensorflow/core/common_runtime/sycl/sycl_util.h"
 #endif  // TENSORFLOW_USE_SYCL
 
+#if GOOGLE_CUDA
+#include "tensorflow/core/kernels/training_ali_ops_gpu.h"
+#endif  // GOOGLE_CUDA
+
 namespace tensorflow {
 
 using CPUDevice = Eigen::ThreadPoolDevice;
@@ -169,7 +173,6 @@ TF_CALL_float(REGISTER_CPU_KERNELS);
 #undef REGISTER_KERNELS
 
 #if GOOGLE_CUDA
-#if CUDA_ATOMIC
 template <class K, class V>
 Status GetInputEmbeddingVarGPU(OpKernelContext* ctx, int input,
                             EmbeddingVarGPU<K, V>** var) {
@@ -288,7 +291,6 @@ TF_CALL_float(REGISTER_GPU_KERNELS);
 
 #undef REGISTER_GPU_KERNELS
 #undef REGISTER_KERNELS
-#endif  // CUDA_ATOMIC
 #endif  // GOOGLE_CUDA
 
 // Note, this op works on cpu only.
@@ -496,7 +498,6 @@ TF_CALL_float(REGISTER_CPU_KERNELS);
 #undef REGISTER_KERNELS
 
 #if GOOGLE_CUDA
-#if CUDA_ATOMIC
 template <typename Device, typename TKey, typename T, bool has_l2_shrinkage>
 class KvSparseApplyFtrlOpGPU : public OpKernel {
  public:
@@ -672,7 +673,6 @@ TF_CALL_float(REGISTER_CPU_KERNELS);
 
 #undef REGISTER_CPU_KERNELS
 #undef REGISTER_KERNELS
-#endif  // CUDA_ATOMIC
 #endif  // GOOGLE_CUDA
 
 // Note, this op works on cpu only.
